Organize files & fix component
