--- src/components/NewsTopStories.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-import SectionTitle from './SectionTitle';
-
-function NewsTopStories() {
-  return (
-    <SectionTitle subtitle="MONTHLY STORIES">Our Monthly Stories.</SectionTitle>
-  );
-}
-
-export default NewsTopStories;
