+++ src/pages/news.js
@@ -1,8 +1,7 @@
 import React from 'react';
 
 import NewsHeader from '../components/NewsHeader';
-import NewsTopStories from '../components/NewsTopStories';
-import MonthStory from '../components/MonthStory';
+import MonthlyStories from '../components/MonthlyStories';
 import NewsCards from '../components/NewsCards';
 import AllStoriesButton from '../components/AllStoriesButton';
 import VideoAudio from '../components/VideoAudioStories';
@@ -15,8 +14,7 @@ function News() {
   return (
     <Page>
       <NewsHeader />
-      <NewsTopStories />
-      <MonthStory />
+      <MonthlyStories />
       <NewsCards />
       <AllStoriesButton />
       <VideoAudio />
